Add delete method to BST
For the delete method, two helper functions had to be created. Considering that if the node to be deleted has two children, a helper function had to be created that searches for a replacement node in the tree. The second helper function, on the other hand, is the recursive delete function itself, which performs the delete searches recursively. In the delete method, the root becomes the return value of the recursive delete function.